requirements.txt
asgiref==3.7.2
asttokens==2.4.1
dj-database-url==2.2.0
Django==5.0.3
django-allauth==0.61.1
django-cors-headers==4.3.1
django-rest-auth==0.9.5
django-rest-authtoken==2.1.4
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
gunicorn==22.0.0
Jinja2==3.1.2
Pillow==10.1.0
#psycopg2==2.9.9
psycopg2-binary==2.9.9
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
uri-template==1.3.0
uritemplate==4.1.1